Engine refactor (#101)
* (WIP) Reworking Memory Allocation system

- created Interface for all Allocators
- getting rid of old implementation from code (20%)
- splitted allocators into different files
- general allocator stores all it's pointers inside allocated block

* Remove leftovers after last memory allocation system implementation

* [WIP] preparation for adding frame debugs implemented on stack

* Huge ZenEngine refactor

There was a lot of bad design decisions made. Fixing it, refactoring everything that needed change. I will need to establish better engine design before going forward

* Fix include statements with project's header files

* Clean up

* Fix nullptr on appInstance

* Remove redundant logging scheduling logic

* Fix wrong engine update time

Unfinished frame was taken into account. Needed to divide finished frames in PerformanceData structure with frames that are "In Progress"

* Improve performance logging readability

+++ Development/Include/Engine/Engine/Core/ISystem.h
@@ -15,7 +15,7 @@ namespace Engine::Core
 		ISystem() {}
 		virtual void StartUp() = 0;
 
-		virtual void Update(const uint32 deltaTime) {};
+		virtual void Update(const double deltaTime) {};
 		virtual void Render() {};
 
 		virtual void ShutDown() = 0;
