+++ Development/Include/Game/Game.h
@@ -5,12 +5,12 @@
 
 #pragma once
 
-#include <Engine/IEngineApplication.h>
+#include "Engine/IEngineApplication.h"
 
 class Game : public IEngineApplication
 {
 public:
 	void StartUp() override;
-	void Update(const uint32 deltaTime) override;
+	void Update(const double deltaTime) override;
 	void ShutDown() override;
 };
